Split pipelines code from app code
