+++ .gitignore
@@ -151,3 +151,4 @@ src/test_gdrive.py
 sandbox
 
 temp*
+*/data/*
