+++ pipelines/codab_utils.py
@@ -1,0 +1,17 @@
+from pipelines import blob_utils
+
+
+def get_blob_name(iso3: str):
+    iso3 = iso3.lower()
+    return f"{blob_utils.PROJECT_PREFIX}/raw/codab/{iso3}.shp.zip"
+
+
+def load_codab_from_blob(iso3: str, admin_level: int = 0):
+    iso3 = iso3.lower()
+    shapefile = f"{iso3}_adm{admin_level}.shp"
+    gdf = blob_utils.load_gdf_from_blob(
+        blob_name=get_blob_name(iso3),
+        shapefile=shapefile,
+        stage="dev",
+    )
+    return gdf
